+++ uui/components/datePickers/DatePickerBody.tsx
@@ -21,10 +21,10 @@ export interface DatePickerBodyProps extends CommonDatePickerBodyProps, IControl
      */
     isHoliday?: (day: Dayjs) => boolean;
     /**
-     * Controls the initially displayed month and year when the picker opens (format: 'YYYY-MM').
+     * Controls the initially displayed month and year when the picker opens (format: 'YYYY-MM-DD').
      * If not specified, defaults to the current month.
      */
-    preselectedViewDate?: string;
+    initialViewDate?: string;
 }
 
 export const uuiDatePickerBody = {
@@ -34,14 +34,14 @@ export const uuiDatePickerBody = {
 export const DatePickerBody = forwardRef(DatePickerBodyComp);
 
 function DatePickerBodyComp(props: DatePickerBodyProps, ref: React.ForwardedRef<HTMLDivElement>) {
-    const { value, onValueChange, preselectedViewDate: _preselectedViewDate = '' } = props;
-    const preselectedViewDate = value || _preselectedViewDate || '';
-    const [month, setMonth] = useState<Dayjs>(getNewMonth(preselectedViewDate));
+    const { value, onValueChange, initialViewDate: _initialViewDate = '' } = props;
+    const initialViewDate = value || _initialViewDate || '';
+    const [month, setMonth] = useState<Dayjs>(getNewMonth(initialViewDate));
     const [view, setView] = useState<ViewType>('DAY_SELECTION');
 
     // sync updated props with internal state
     useEffect(() => {
-        setMonth(getNewMonth(preselectedViewDate));
+        setMonth(getNewMonth(initialViewDate));
         setView('DAY_SELECTION');
     }, [value, setMonth]);