+++ changelog.md
@@ -10,7 +10,8 @@
 * [Pagination]: pagination elements are not grouped into a list ([#2942](https://github.com/epam/UUI/pull/2942))
   * Add "Previous page" and "Next page" `aria-label`-s to the previous and next pages respectively
 * [DatePickers]: fixed calendar matrix calculation to use locale-specific `firstDayOfWeek` from dayjs instead of hardcoded i18n configuration, ensuring correct calendar layout for any locale.
-* [RangeDatePicker}: added option to display clear icon when at least one date is selected
+* [RangeDatePicker]: added option to display clear icon when at least one date is selected
+* [DatePicker], [RangeDatePicker]: added ability to set initial Date in calendar body via `initialViewDate` prop
 
 
 # 6.3.1 - 08.10.2025